🛑 Don't allow the removal of the filesystem's '.' and '..'.
FossilOrigin-Name: 738c7480e3b68601ba9231138fc0f3e9ffcf943f1f8aea3e60dfc5ae32b7716a

+++ src/db_module/sqlite.rs
@@ -274,8 +274,10 @@ fn get_dentry_single_at_time(
 }
 
 fn delete_dentry_local(parent: u32, name: &str, tx: &Connection) -> Result<()> {
+    if(!(name==".." || name == ".")){
     let sql = "DELETE FROM dentry WHERE parent_id=$1 and name=$2";
     tx.execute(sql, params![parent, name])?;
+    }
     Ok(())
 }
 
@@ -375,10 +377,10 @@ impl Sqlite {
         let conn = Connection::open(path)?;
         // enable foreign key. Sqlite ignores foreign key by default.
         conn.execute("PRAGMA foreign_keys=ON", NO_PARAMS)?;
-        conn.execute("CREATE TABLE tdentry AS SELECT * FROM (SELECT * FROM (select * from dentry_audit WHERE timestamp_utc < (?1) ORDER BY timestamp_utc DESC) GROUP BY name) as t WHERE TG_OP <> 'DELETE';", params![time])?;
-        conn.execute("CREATE TABLE tmetadata AS SELECT * FROM (SELECT * FROM (select * from metadata_audit WHERE timestamp_utc < (?1) ORDER BY timestamp_utc DESC) GROUP BY id) as t WHERE TG_OP <> 'DELETE';", params![time])?;
-        conn.execute("CREATE TABLE tdata AS SELECT * FROM (SELECT * FROM (select * from data_audit WHERE timestamp_utc < (?1) ORDER BY timestamp_utc DESC) GROUP BY file_id) as t WHERE TG_OP <> 'DELETE';", params![time])?;
-        conn.execute("CREATE TABLE txattr AS SELECT * FROM (SELECT * FROM (select * from xattr_audit WHERE timestamp_utc < (?1) ORDER BY timestamp_utc DESC) GROUP BY file_id, name) as t WHERE TG_OP <> 'DELETE';", params![time] )?;
+        conn.execute("CREATE TEMP TABLE tdentry AS SELECT * FROM (SELECT * FROM (select * from dentry_audit WHERE timestamp_utc < (?1) ORDER BY timestamp_utc DESC) GROUP BY name) as t WHERE TG_OP <> 'DELETE';", params![time])?;
+        conn.execute("CREATE TEMP TABLE tmetadata AS SELECT * FROM (SELECT * FROM (select * from metadata_audit WHERE timestamp_utc < (?1) ORDER BY timestamp_utc DESC) GROUP BY id) as t WHERE TG_OP <> 'DELETE';", params![time])?;
+        conn.execute("CREATE TEMP TABLE tdata AS SELECT * FROM (SELECT * FROM (select * from data_audit WHERE timestamp_utc < (?1) ORDER BY timestamp_utc DESC) GROUP BY file_id) as t WHERE TG_OP <> 'DELETE';", params![time])?;
+        conn.execute("CREATE TEMP TABLE txattr AS SELECT * FROM (SELECT * FROM (select * from xattr_audit WHERE timestamp_utc < (?1) ORDER BY timestamp_utc DESC) GROUP BY file_id, name) as t WHERE TG_OP <> 'DELETE';", params![time] )?;
         Ok(Sqlite { conn })
     }
     pub fn new_read_only(path: &Path) -> Result<Self> {
